fix: manually extract path parameters in panels (#1988)
Due to #1978, `<Panels />` (rendered by `<BaseLayout />`) is no longer
rendered within the sub-routers for the model detail pages. This means
that the panels no longer inherit the route, and so `useParams` doesn't
return the correct parameters.

This fix manually uses `useMatch` provided by react-router to
re-parse the location with a manually-specified path. An alternate fix
could involve re-structuring the render tree, to ensure that `<Panels
/>` is again rendered as a child of the router.