test: update run playwright action and keycloak backend e2e test fix (#2059)
* ci: set correct working directory  for run-playwright action

* ci: remove setup-node cache

Using cache-dependency-path comes with the caveat that no relative paths are allowed other you get the following error,

"Error: Invalid pattern '<path>'. Relative pathing '.' and '..' is not allowed.

* ci: specify node-version

* test: remove CLI login for keycloak user creation

* ci: use a normalised path

Tests successfully passed but the upload-artifact step failed because it doesn't accept a relative path so normalise first and use that value.

* ci: remove node-version input

* ci: allow workflow to specify dashboard address scheme

When running the dashboards tests from the jimm repo, the dashboard is served via https, so allow the workflow to specify the scheme rather than hardcoding it in the action.

* ci: add scheme for automatically determined dashboard URL

* ci: use a fixed version of jimm in CI tests

Using a fixed version of JIMM in CI means that you are not guessing what broke when you made a change, jimm or the dashboard.
Then when a new feature is released in jimm that the dashboard wants to take advantage of the new feature.